skip over SOFT line parse error; organize; use logging
